Now requiring Java 1.5
The goal was to run on Java 1.2, but that decision really restricted the
development and made the whole experience not very enjoyable. Moving to
JDK 1.5 means we have access to a lot more stuff in the core language
and we are still capable of unofficially targetting Windows 95, and
officially targetting Mac OS X 10.4 Tiger.